requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
redis==4.6.0
openai==1.3.0
pdfplumber==0.10.2
python-docx==0.8.11
docx2txt==0.8
langchain==0.0.335
nltk==3.8.1
slowapi==0.1.8
gunicorn==21.2.0
python-dotenv==1.0.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1